Change random number to a float

+++ src/main.cpp
@@ -300,7 +300,7 @@ void drawPlayer(Player * player) {
 void spawnEnemy() {
     // Spawn the enemies at [48, -48] on the x-coordinate
     int offset = -48;
-    int randomNumber = (rand() % 97) + offset;
+    float randomNumber = (rand() % 97) + offset;
     enemies.push_back(Enemy((Vector3){randomNumber, 0.0f, 40.0f}));
 }
